+++ app/Http/Controllers/ProductController.php
@@ -221,9 +221,11 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        $product->delete();
-        return redirect()
-            ->back()
-            ->with('success', 'Produk berhasil dihapus');
+        if (Utils::isAdmin()) {
+            $product->delete();
+            return redirect()
+                ->back()
+                ->with('success', 'Produk berhasil dihapus');
+        }
     }
 }
